Added Gallery Of Images | Added Post request with Rating

+++ app/src/main/java/com/cavigna/movieapp/model/models/model/images/Backdrop.kt
@@ -13,7 +13,7 @@ data class Backdrop(
     @SerializedName("file_path")
     var filePath: String = "",
     @SerializedName("vote_average")
-    var voteAverage: Int = 0,
+    var voteAverage: Double = 0.0,
     @SerializedName("vote_count")
     var voteCount: Int = 0,
     @SerializedName("width")
